# scooter-rental/src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Scooter Rental API
  description: REST API for scooter rental system
  version: 1.0.0
  contact:
    name: Albert
    email: albertforeber2@gmail.com
servers:
  - url: 'http://localhost:8080/api'
    description: Local development
security:
  - bearerAuth: [ ]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  examples:
    JsonParseException:
      summary: "Malformed JSON"
      description: "Appears when JSON is malformed"
      value:
        status: 400
        error: "Parse exception"
        message: "Failed to parse your request. Check all fields are correct"
        details:
          - "Request JSON body is malformed"
    MismatchedInputException:
      summary: "Invalid field or field is excessive"
      description: "Appears when wrong type given (e.g. array instead of single value) in JSON request or field is exessive"
      value:
        status: 400
        error: "Parse exception"
        message: "Failed to parse your request. Check all fields are correct"
        details:
          - "Invalid type for field or field is excessive: not_expected_field_name"
    InvalidFieldValue:
      summary: "Invalid field value"
      description: "Appears when field value is wrong (e.g. expected enum, but given value doesn't match any of expected values)"
      value:
        status: 400
        error: "Parse exception"
        message: "Failed to parse your request. Check all fields are correct"
        details:
          - "Invalid value for field: status"
    BodyValidationException:
      summary: "Body validation exception"
      description: "Appears when some of request body field failed to validate (e.g. field expected to be filled but sent null)"
      value:
        status: 400
        error: "Validation exception"
        message: "Some of fields failed to validate"
        details:
          - "name: Field 'name' should not be empty"
    SimpleValidationException:
      summary: "Param validation exception"
      description: "Appears when some of path variables or params failed to validate (e.g. param expected to be positive but sent negative)"
      value:
        status: 400
        error: "Validation exception"
        message: "Some of path variables or params failed to validate"
        details:
          - "page: Param 'page' must be positive number"
    TypeMismatchException:
      summary: "Wrong parameter type"
      description: "Appears when path variable or request param cannot be converted (e.g. non-numeric id)"
      value:
        status: 400
        error: "Type mismatch"
        message: "Some of given parameters are of the wrong type"
        details:
          - "Wrong value for parameter 'id': abc"
    NoSuchEntityException:
      summary: "No such entity"
      description: "Appears when user requested entity that doesn't exist"
      value:
        status: 400
        error: "No such entity exception"
        message: "Requested data was not found"
        details:
          - "No tariff found with id: 1"
    UnavailableActionException:
      summary: "Unavailable action"
      description: "Appears when user tries to perform unavailable action (e.g rent an occupied scooter)"
      value:
        status: 400
        error: "Unavailable action exception"
        message: "You are not able to perform such action"
        details:
          - "Forbidden to rent occupied scooter"
    InvalidToken:
      summary: "Invalid token"
      description: "Appears provided invalid token (e.g expired)"
      value:
        status: 401
        error: "Invalid token"
        message: "Error during token validation"
        details:
          - "JWT expired"
    Unauthorized:
      summary: "No access token"
      description: "Appears when endpoints requires authorization but token was not provided"
      value:
        status: 401
        error: "Unauthorized"
        message: "No access token found"
...
    AccessDeniedException:
      summary: "Access denied"
      description: "Appears when user doesn't have enough permissions to user this endpoint"
      value:
        status: 403
        error: "Access denied"
        message: "You don't have required permissions"
    NoRequiredEntityException:
      summary: "No required entity"
      description: "Appears when request required some entity but server failed to find it (e.g. default tariff not found)"
      value:
        status: 500
        error: "No required entity exception"
        message: "Some of required data was not found. Please contact support team"
    DataManipulationException:
      summary: "Data manipulation exception"
      description: "Appears when server failed to process required data (e.g. database failed to perform some actions)"
      value:
        status: 500
        error: "Data manipulation exception"
        message: "Something went wrong during processing your request, please try later"
    Exception:
      summary: "Interval server error"
      description: "Appears when unexpected error occurred on server side"
      value:
        status: 500
        error: "Internal server error"
        message: "Something went wrong, please try again later"
  responses:
    Error400Full:
      description: "User side errors"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            JsonParseException:
              $ref: '#/components/examples/JsonParseException'
            MismatchedInputException:
              $ref: '#/components/examples/MismatchedInputException'
            InvalidFieldValue:
              $ref: '#/components/examples/InvalidFieldValue'
            BodyValidationException:
              $ref: '#/components/examples/BodyValidationException'
            SimpleValidationException:
              $ref: '#/components/examples/SimpleValidationException'
            TypeMismatchException:
              $ref: '#/components/examples/TypeMismatchException'
            NoSuchEntityException:
              $ref: '#/components/examples/NoSuchEntityException'
            UnavailableActionException:
              $ref: '#/components/examples/UnavailableActionException'
...
    Error401:
      description: "Auth errors"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            InvalidToken:
              $ref: '#/components/examples/InvalidToken'
            Unauthorized:
              $ref: '#/components/examples/Unauthorized'
...
    Error403:
      description: "Permission error"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            AccessDeniedException:
              $ref: '#/components/examples/AccessDeniedException'
    Error500:
      description: "Server side errors"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            NoRequiredEntityException:
              $ref: '#/components/examples/NoRequiredEntityException'
            DataManipulationException:
              $ref: '#/components/examples/DataManipulationException'
            Exception:
              $ref: '#/components/examples/Exception'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Validation exception
        message:
          type: string
          example: Some of fields failed to validate
        details:
          type: array
          items:
            type: string
          example:
            - "Field 'email' must contain well-formed email address"
...
    TariffConciseResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Per minute
...
    SubscriptionTariffResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Per month
        basePrice:
          type: number
          example: 250.50
        durationDays:
          type: integer
          example: 1
    CurrentSubscriptionResponse:
      type: object
      properties:
        tariff:
          $ref: '#/components/schemas/SubscriptionTariffResponse'
        nextBillingDate:
          type: string
          format: date
          example: 2026-06-13
    SubscribedResponse:
      type: object
      properties:
        tariff:
          $ref: '#/components/schemas/TariffConciseResponse'
        nextBillingDate:
          nextBillingDate:
            type: string
            format: date
            example: 2026-06-13
        instantPay:
          type: number
          example: 250.50
    CreateSubscriptionTariffRequest:
      type: object
      required: [ name, basePrice, durationDays ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        basePrice:
          type: number
          minimum: 1
          example: 250.50
        durationDays:
          type: integer
          minimum: 1
    UpdateSubscriptionTariffRequest:
      type: object
      properties:
        durationDays:
          type: integer
          minimum: 1

# --------------- BILLING ---------------
...
paths:
...
  /subscriptions:
    get:
      tags: [ Subscriptions ]
      summary: "List subscription tariffs or admin view of user subscription"
      description: |
        Without `user_id` — all subscriptions (`SCOPE_tariff:view`).
        With `user_id` — current user's subscription (`SCOPE_tariff:view_admin`).
      parameters:
        - name: pageSize
          in: query
          schema: { type: integer, default: 10, minimum: 1 }
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: user_id
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/TariffConciseResponse'
                  - $ref: '#/components/schemas/CurrentSubscriptionResponse'
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      tags: [ Subscriptions ]
      summary: "Create subscription tariff"
      description: "`SCOPE_tariff:manage`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionTariffRequest'
      responses:
        '201':
          description: "Created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionTariffResponse'
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /subscriptions/my:
    get:
      tags: [ Subscriptions ]
      summary: "My current subscription"
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentSubscriptionResponse'
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /subscriptions/unsubscribe:
    delete:
      tags: [ Subscriptions ]
      summary: "Cancel subscription"
      description: "Required `SCOPE_tariff:subscribe`, unavailable during ongoing trips"
      responses:
        '204':
          description: "Unsubscribed"
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /subscriptions/{id}:
    get:
      tags: [ Subscriptions ]
      summary: "Subscription tariff by id"
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionTariffResponse'
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
    patch:
      tags: [ Subscriptions ]
      summary: "Update subscription tariff duration"
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionTariffRequest'
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionTariffResponse'
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags: [ Subscriptions ]
      summary: "Delete subscription tariff"
      description: "Force is used to refund all subscriptions and force delete`."
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: force
          in: query
          schema: { type: boolean, default: false }
      responses:
        '204':
          description: "Deleted"
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /subscriptions/{id}/subscribe:
    post:
      tags: [ Subscriptions ]
      summary: "Subscribe to tariff"
      description: "Required `SCOPE_tariff:subscribe`"
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '201':
          description: "Subscribed"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscribedResponse'
        '400':
          $ref: '#/components/responses/Error400Full'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'